Build[able] sources after: -Build NDK libs; -Android tools\Add Support Library; -Android tools\Fix project properties; -Source\Clean Up -Ant release

+++ src/org/geometerplus/fbreader/network/atom/ATOMDateConstruct.java
@@ -109,8 +109,8 @@ public abstract class ATOMDateConstruct extends ATOMCommonAttributes implements 
 			if (!Character.isDigit(ch)) {
 				return false;
 			}
-			num = 10 * num + ((int) (ch - '0'));
-			fnum += fmult * ((int) (ch - '0'));
+			num = 10 * num + (ch - '0');
+			fnum += fmult * (ch - '0');
 			fmult *= 0.1f;
 			if (start == end) {
 				switch (log) {
